all examples work
