--- examples/new_syntax.mpp
@@ -1,64 +1,0 @@
-function f (num a) {
-    print(a + 1)
-}
-
-num a = 12
-a += 3.5
-print(a)
-str x = "Hello, World!"
-print(x)
-printflush(message1)
-
-f(9)
-
-# read result cell1 0
-# write result cell1 0
-# draw clear 0 0 0 0 0 0
-# print "frog"
-# drawflush display1
-# printflush message1
-# getlink result 0
-# control enabled block1 0 0 0 0
-# radar enemy any any distance turret1 1 result
-# sensor result block1 @copper
-# set result 0
-# op add result a b
-# wait 0.5
-# lookup item result 0
-# packcolor result 1 0 0 1
-# end
-# jump -1 notEqual x false
-# ubind @poly
-# ucontrol move 0 0 0 0 0
-# uradar enemy any any distance 0 1 result
-# ulocate building core true @copper outx outy found building
-
-# control enabled block1 e 0 0 0
-# control shoot block1 x y s 0
-# control shootp block1 u s 0 0
-# control config block1 c 0 0 0
-# control color block1 c 0 0 0
-
-# ucontrol idle 0 0 0 0 0
-# ucontrol stop 0 0 0 0 0
-# ucontrol move x y 0 0 0
-# ucontrol approach x y r 0 0
-# ucontrol boost e 0 0 0 0
-# ucontrol target x y s 0 0
-# ucontrol targetp u s 0 0 0
-# ucontrol itemDrop t a 0 0 0
-# ucontrol itemTake f i a 0 0
-# ucontrol payDrop 0 0 0 0 0
-# ucontrol payTake t 0 0 0 0
-# ucontrol payEnter 0 0 0 0 0
-# ucontrol mine x y 0 0 0
-# ucontrol flag v 0 0 0 0
-# ucontrol build x y b r c
-# ucontrol getBlock x y t b 0
-# ucontrol within x y r r 0
-# ucontrol unbind 0 0 0 0 0
-
-# ulocate ore core true @copper outx outy found building
-# ulocate building core true @copper outx outy found building
-# ulocate spawn core true @copper outx outy found building
-# ulocate damaged core true @copper outx outy found building
