+++ examples/noise.mpp
@@ -7,10 +7,10 @@ while (true) {
     print(seed)
     printflush(message1)
 
-    for (y = 0; y < dsize; y += 1) {
-        for (x = 0; x < dsize; x += 1) {
-            num val = op(noise, x + seed, y + seed)
-            val = op(abs, val, _)
+    for (num y = 0; y < dsize; y += 1) {
+        for (num x = 0; x < dsize; x += 1) {
+            num val = noise(x + seed, y + seed)
+            val = abs(val)
             val *= 255
             draw.color(val, val, val, 255)
             draw.rect(x * PM, y * PM, PM, PM)
