+++ examples/operators.mpp
@@ -1,5 +1,5 @@
 num a = 1 + 2 - 3 * 4 // 5 ** 6 % 7
-num b = 10 << 3 >> (1 + 2)
+num b = 10 << 3 >> a
 b %= 3
 b ^= 127
 b &= 32
