+++ mlogpp/lexer.py
@@ -4,7 +4,7 @@ import string
 from .preprocess import Preprocessor
 from .error import Error
 from .util import Position, sanitize
-from .node_rewrite import NativeCallNode
+from .node import NativeCallNode
 from .tokens import *
