+++ mlogpp/optimizer.py
@@ -2,6 +2,7 @@ import re
 
 from .functions import PRECALC, JC_REPLACE
 from .instruction import *
+from .value import *
 
 
 class Optimizer:
@@ -17,7 +18,7 @@ class Optimizer:
 
         for _ in range(10):
             for i in range(1, 10):
-                while (cf := Optimizer._single_use_tmp(code, i))[1]:
+                while (cf := Optimizer._single_use(code, i))[1]:
                     code = cf[0]
                 while (cf := Optimizer._precalculate_optimize(code))[1]:
                     code = cf[0]
@@ -29,7 +30,7 @@ class Optimizer:
         return Instructions([ins for ins in code.iter() if not isinstance(ins, NoopInstruction)])
 
     @staticmethod
-    def _single_use_tmp(code: Instructions, forward: int = 1) -> tuple[Instructions, bool]:
+    def _single_use(code: Instructions, forward: int = 1) -> tuple[Instructions, bool]:
         """
         optimize single use temporary variables
         """
@@ -39,6 +40,9 @@ class Optimizer:
         uses = {}
         for ins in code.iter():
             for var in ins.variables():
+                if var.startswith("@"):
+                    continue
+
                 if var not in uses:
                     uses[var] = 0
                 uses[var] += 1
@@ -52,36 +56,36 @@ class Optimizer:
                     cfi = code[fi]
                     match ins:
                         case MInstruction(MInstructionType.SET, [name, value]):
-                            if name.startswith("__tmp") and uses.get(name, 0) == 2:
+                            if name.startswith("@") or value.startswith("@"):
+                                continue
+
+                            if uses.get(name, 0) == 1:
+                                code[i] = NoopInstruction()
+                                found = True
+
+                            elif uses.get(name, 0) == 2:
                                 if name in cfi.variables():
                                     code[i] = NoopInstruction()
                                     cfi.param_replace(name, value)
                                     found = True
 
-                            elif name.startswith("__f_lvar") and value.startswith("__f_a"):
-                                code[i] = NoopInstruction()
-                                code.param_replace(name, value)
-                                found = True
-
                         case MInstruction(MInstructionType.GETLINK, [name, _]):
-                            if isinstance(cfi, MInstruction) and cfi.type == MInstructionType.SET and uses.get(name,
-                                                                                                               0) == 2 and \
-                                    name.startswith("__tmp") and cfi.params[1] == name:
+                            if isinstance(cfi, MInstruction) and cfi.type == MInstructionType.SET and \
+                                    uses.get(name, 0) == 2 and name.startswith("__tmp") and cfi.params[1] == name:
                                 ins.params[0] = cfi.params[0]
                                 code[fi] = NoopInstruction()
                                 found = True
 
                         case MInstruction(MInstructionType.OP, [op, result, op1, op2]):
-                            if isinstance(cfi, MInstruction) and cfi.type == MInstructionType.SET and uses.get(result,
-                                                                                                               0) == 2 and \
-                                    result.startswith("__tmp") and cfi.params[1] == result:
+                            if isinstance(cfi, MInstruction) and cfi.type == MInstructionType.SET and \
+                                    uses.get(result, 0) == 2 and result.startswith("__tmp") and cfi.params[1] == result:
                                 ins.params[1] = cfi.params[0]
                                 code[fi] = NoopInstruction()
                                 found = True
 
                             if isinstance(cfi, MppInstructionOJump) and uses.get(result, 0) == 2 and \
                                     result.startswith("__tmp") and cfi.op1 == result and \
-                                    cfi.op == "notEqual" and cfi.op2 == "true":
+                                    cfi.op == "equal" and cfi.op2 in ("0", "false"):
 
                                 if op in JC_REPLACE:
                                     code[i] = NoopInstruction()
@@ -91,9 +95,8 @@ class Optimizer:
                                     found = True
 
                         case MInstruction(MInstructionType.SENSOR, [result, _, _]):
-                            if isinstance(cfi, MInstruction) and cfi.type == MInstructionType.SET and uses.get(result,
-                                                                                                               0) == 2 and \
-                                    result.startswith("__tmp") and cfi.params[1] == result:
+                            if isinstance(cfi, MInstruction) and cfi.type == MInstructionType.SET and \
+                                    uses.get(result, 0) == 2 and result.startswith("__tmp") and cfi.params[1] == result:
                                 ins.params[0] = cfi.params[0]
                                 code[fi] = NoopInstruction()
                                 found = True
@@ -101,18 +104,15 @@ class Optimizer:
                         case MInstruction(MInstructionType.URADAR, [_, _, _, _, _, _, result]) | \
                                 MInstruction(MInstructionType.RADAR, [_, _, _, _, _, _, result]):
 
-                            if isinstance(cfi, MInstruction) and cfi.type == MInstructionType.SET and uses.get(result,
-                                                                                                               0) == 2 and \
-                                    result.startswith("__tmp") and cfi.params[1] == result:
+                            if isinstance(cfi, MInstruction) and cfi.type == MInstructionType.SET and \
+                                    uses.get(result, 0) == 2 and result.startswith("__tmp") and cfi.params[1] == result:
                                 ins.params[6] = cfi.params[0]
                                 code[fi] = NoopInstruction()
                                 found = True
 
                         case MInstruction(MInstructionType.READ, [result, _, _]):
-                            if isinstance(cfi, MInstruction) and cfi.type == MInstructionType.SET and uses.get(result,
-                                                                                                               0) == 2 and \
-                                    result.startswith("__tmp") and cfi.params[1] == result:
-
+                            if isinstance(cfi, MInstruction) and cfi.type == MInstructionType.SET and \
+                                    uses.get(result, 0) == 2 and result.startswith("__tmp") and cfi.params[1] == result:
                                 ins.params[0] = cfi.params[0]
                                 code[fi] = NoopInstruction()
                                 found = True
